tools/ffl-cli/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telnyx-ffl-cli"
version = "0.1.0"
description = "CLI for reporting friction when using Telnyx APIs (Friction Feedback Loop)"
readme = "README.md"
authors = [
    {name = "Telnyx", email = "ai-fde@telnyx.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/team-telnyx/aifde-ffl-cli"
Repository = "https://github.com/team-telnyx/aifde-ffl-cli.git"
Issues = "https://github.com/team-telnyx/aifde-ffl-cli/issues"

[project.scripts]
friction-report = "telnyx_ffl_cli.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
